Addressed requested changes - rename var - move files to root http_api
