+++ framework/src/modules/chain/helpers/http_api.js
@@ -33,7 +33,7 @@ require('./swagger').getValidator();
  * @see Parent: {@link helpers}
  * @requires extend
  * @requires lodash
- * @requires helpers/api_codes
+ * @requires api_codes
  * @requires helpers/check_ip_in_list
  * @property {Object} middleware
  * @todo Add description for the module and the properties
