+++ framework/test/mocha/functional/http/post/0.transfer.js
@@ -23,7 +23,7 @@ const phases = require('../../../common/phases');
 const sendTransactionPromise = require('../../../common/helpers/api')
 	.sendTransactionPromise;
 const randomUtil = require('../../../common/utils/random');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const Bignum = require('../../../../../src/modules/chain/helpers/bignum.js');
 
 const { NORMALIZER } = global.constants;
@@ -69,7 +69,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Failed to verify signature');
 				badTransactions.push(transaction);
@@ -82,7 +82,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Invalid transaction id');
 				badTransactions.push(transaction);
@@ -98,7 +98,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Invalid transaction amount');
 				badTransactions.push(transaction);
@@ -114,7 +114,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Account does not have enough LSK: ${account.address} balance: 0`
@@ -132,7 +132,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.match(
 					/^Account does not have enough LSK: [0-9]+L balance: /
@@ -159,7 +159,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				signedTransactionFromGenesis,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid sender. Can not send from genesis account'
@@ -178,7 +178,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 		it('sending transaction with same id twice should fail', async () => {
 			return sendTransactionPromise(
 				goodTransaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction is already processed: ${goodTransaction.id}`
@@ -191,7 +191,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				cloneGoodTransaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction is already processed: ${cloneGoodTransaction.id}`
@@ -204,7 +204,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 			return sendTransactionPromise(
 				cloneGoodTransaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction is already processed: ${cloneGoodTransaction.id}`
@@ -237,7 +237,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction timestamp. Timestamp is in the future'
@@ -265,7 +265,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 
 						return sendTransactionPromise(
 							transaction,
-							errorCodes.PROCESSING_ERROR
+							apiCodes.PROCESSING_ERROR
 						).then(res => {
 							expect(res.body.message).to.not.be.empty;
 							badTransactions.push(transaction);
@@ -347,7 +347,7 @@ describe('POST /api/transactions (type 0) transfer funds', () => {
 		it('sending already confirmed transaction should fail', async () => {
 			return sendTransactionPromise(
 				goodTransaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction is already confirmed: ${goodTransaction.id}`
