+++ framework/test/mocha/functional/http/post/1.second_secret.js
@@ -22,7 +22,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../common/helpers/api');
 const randomUtil = require('../../../common/utils/random');
 const waitFor = require('../../../common/utils/wait_for');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 
 const { FEES, NORMALIZER } = global.constants;
@@ -89,7 +89,7 @@ describe('POST /api/transactions (type 1) register second passphrase', () => {
 			});
 
 			return apiHelpers
-				.sendTransactionPromise(transaction, errorCodes.PROCESSING_ERROR)
+				.sendTransactionPromise(transaction, apiCodes.PROCESSING_ERROR)
 				.then(res => {
 					expect(res.body.message).to.be.equal(
 						'Sender does not have a second signature'
@@ -105,7 +105,7 @@ describe('POST /api/transactions (type 1) register second passphrase', () => {
 			});
 
 			return apiHelpers
-				.sendTransactionPromise(transaction, errorCodes.PROCESSING_ERROR)
+				.sendTransactionPromise(transaction, apiCodes.PROCESSING_ERROR)
 				.then(res => {
 					expect(res.body.message).to.be.equal(
 						`Account does not have enough LSK: ${
