+++ framework/test/mocha/functional/http/post/2.delegate.js
@@ -22,7 +22,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../common/helpers/api');
 const randomUtil = require('../../../common/utils/random');
 const waitFor = require('../../../common/utils/wait_for');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 
 const { FEES, NORMALIZER } = global.constants;
@@ -98,7 +98,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Account does not have enough LSK: ${
@@ -129,7 +129,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Username is undefined');
 				badTransactions.push(transaction);
@@ -145,7 +145,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Invalid transaction body - Failed to validate delegate schema: Object didn't pass validation for format username: ${username}`
@@ -163,7 +163,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Username is too long. Maximum is 20 characters'
@@ -180,7 +180,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Username must be lowercase');
 				badTransactions.push(transaction);
@@ -213,7 +213,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Account is already a delegate');
 				badTransactionsEnforcement.push(transaction);
@@ -228,7 +228,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Username ${account.username} already exists`
@@ -245,7 +245,7 @@ describe('POST /api/transactions (type 2) register delegate', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal('Account is already a delegate');
 				badTransactionsEnforcement.push(transaction);
