+++ framework/test/mocha/functional/http/post/3.votes.js
@@ -23,7 +23,7 @@ const randomUtil = require('../../../common/utils/random');
 const waitFor = require('../../../common/utils/wait_for');
 const elements = require('../../../common/utils/elements');
 const apiHelpers = require('../../../common/helpers/api');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 
 const {
@@ -267,7 +267,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid vote at index 0 - Invalid vote format'
@@ -291,7 +291,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid vote at index 0 - Invalid vote format'
@@ -313,7 +313,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid vote at index 0 - Invalid vote format'
@@ -335,7 +335,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid vote at index 0 - Invalid vote format'
@@ -357,7 +357,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid vote at index 0 - Invalid vote type'
@@ -375,7 +375,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Account does not have enough LSK: ${
@@ -394,7 +394,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.equal('Delegate not found');
 				badTransactions.push(transaction);
@@ -421,7 +421,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Failed to remove vote, delegate "${
@@ -483,7 +483,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					'Invalid transaction body - Failed to validate vote schema: Array is too long (34), maximum 33'
@@ -551,7 +551,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Failed to add vote, delegate "${
@@ -594,7 +594,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Maximum number of ${ACTIVE_DELEGATES} votes exceeded (1 too many)`
@@ -628,7 +628,7 @@ describe('POST /api/transactions (type 3) votes', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Invalid transaction body - Failed to validate vote schema: Array is too long (${MAX_VOTES_PER_TRANSACTION +
