+++ framework/test/mocha/functional/http/post/5.dapps.js
@@ -23,7 +23,7 @@ const accountFixtures = require('../../../fixtures/accounts');
 const randomUtil = require('../../../common/utils/random');
 const waitFor = require('../../../common/utils/wait_for');
 const apiHelpers = require('../../../common/helpers/api');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 
 const { FEES, NORMALIZER } = global.constants;
@@ -100,7 +100,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/Missing required property: category$/
@@ -118,7 +118,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type integer but found type string'
@@ -136,7 +136,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(/Value -1 is less than minimum 0$/);
 					badTransactions.push(transaction);
@@ -152,7 +152,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/Value 9 is greater than maximum 8$/
@@ -199,7 +199,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type string but found type integer'
@@ -235,7 +235,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/String is too long \(161 chars\), maximum 160$/
@@ -270,7 +270,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type string but found type integer'
@@ -290,7 +290,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal('Invalid application icon link');
 					badTransactions.push(transaction);
@@ -308,7 +308,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						'Invalid application icon file type'
@@ -330,7 +330,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal('Invalid application link');
 					badTransactions.push(transaction);
@@ -346,7 +346,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type string but found type integer'
@@ -366,7 +366,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal('Invalid application file type');
 					badTransactions.push(transaction);
@@ -384,7 +384,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(/Missing required property: name$/);
 					badTransactions.push(transaction);
@@ -400,7 +400,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type string but found type integer'
@@ -418,7 +418,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/String is too short \(0 chars\), minimum 1$/
@@ -439,7 +439,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/String is too long \(33 chars\), maximum 32$/
@@ -474,7 +474,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						'Invalid transaction body - Failed to validate dapp schema: Expected type string but found type integer'
@@ -510,7 +510,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/String is too long \(161 chars\), maximum 160$/
@@ -546,7 +546,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal(
 						`Encountered duplicate tag: ${tag} in application`
@@ -566,7 +566,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(/Missing required property: type$/);
 					badTransactions.push(transaction);
@@ -582,7 +582,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(/Value -1 is less than minimum 0$/);
 					badTransactions.push(transaction);
@@ -598,7 +598,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(/Value -1 is less than minimum 0$/);
 					badTransactions.push(transaction);
@@ -615,7 +615,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.equal('Invalid application type');
 					badTransactions.push(transaction);
@@ -635,7 +635,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.equal(
 					`Application name already exists: ${dapp.name}`
@@ -654,7 +654,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.equal(
 					`Application link already exists: ${dapp.link}`
@@ -671,7 +671,7 @@ describe('POST /api/transactions (type 5) register dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.equal(
 					`Account does not have enough LSK: ${
