+++ framework/test/mocha/functional/http/post/7.dapps.out_transfer.js
@@ -24,7 +24,7 @@ const randomUtil = require('../../../common/utils/random');
 const waitFor = require('../../../common/utils/wait_for');
 const elements = require('../../../common/utils/elements');
 const apiHelpers = require('../../../common/helpers/api');
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const common = require('./common');
 
 const { FEES, NORMALIZER } = global.constants;
@@ -112,7 +112,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Missing required property: dappId'
@@ -133,7 +133,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type integer'
@@ -154,7 +154,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						"Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type number, Object didn't pass validation for format id: 1.2"
@@ -175,7 +175,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type array'
@@ -196,7 +196,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						"Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type object, Object didn't pass validation for format id: {}"
@@ -216,7 +216,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: String is too short (0 chars), minimum 1'
@@ -237,7 +237,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						`Invalid transaction body - Failed to validate outTransfer schema: Object didn't pass validation for format id: ${invalidDappId}`
@@ -260,7 +260,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Missing required property: transactionId'
@@ -281,7 +281,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type integer'
@@ -302,7 +302,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						"Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type number, Object didn't pass validation for format id: 1.2"
@@ -323,7 +323,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type array'
@@ -344,7 +344,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						"Invalid transaction body - Failed to validate outTransfer schema: Expected type string but found type object, Object didn't pass validation for format id: {}"
@@ -364,7 +364,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate outTransfer schema: String is too short (0 chars), minimum 1'
@@ -385,7 +385,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						`Invalid transaction body - Failed to validate outTransfer schema: Object didn't pass validation for format id: ${invalidTransactionId}`
@@ -406,7 +406,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 				);
 				transaction.recipientId = 1;
 
-				return sendTransactionPromise(transaction, errorCodes.BAD_REQUEST).then(
+				return sendTransactionPromise(transaction, apiCodes.BAD_REQUEST).then(
 					async () => {
 						badTransactions.push(transaction);
 					}
@@ -423,7 +423,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 				);
 				transaction.recipientId = 1.2;
 
-				return sendTransactionPromise(transaction, errorCodes.BAD_REQUEST).then(
+				return sendTransactionPromise(transaction, apiCodes.BAD_REQUEST).then(
 					async () => {
 						badTransactions.push(transaction);
 					}
@@ -440,7 +440,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 				);
 				transaction.recipientId = [];
 
-				return sendTransactionPromise(transaction, errorCodes.BAD_REQUEST).then(
+				return sendTransactionPromise(transaction, apiCodes.BAD_REQUEST).then(
 					async () => {
 						badTransactions.push(transaction);
 					}
@@ -457,7 +457,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 				);
 				transaction.recipientId = {};
 
-				return sendTransactionPromise(transaction, errorCodes.BAD_REQUEST).then(
+				return sendTransactionPromise(transaction, apiCodes.BAD_REQUEST).then(
 					async () => {
 						badTransactions.push(transaction);
 					}
@@ -475,7 +475,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(() => {
 					badTransactions.push(transaction);
 				});
@@ -491,7 +491,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 					account.passphrase
 				);
 
-				return sendTransactionPromise(transaction, errorCodes.BAD_REQUEST).then(
+				return sendTransactionPromise(transaction, apiCodes.BAD_REQUEST).then(
 					async () => {
 						badTransactions.push(transaction);
 					}
@@ -511,7 +511,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.be.equal(
 						'Invalid transaction body - Failed to validate transaction schema: Value -1 is less than minimum 0'
@@ -542,7 +542,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 						return sendTransactionPromise(
 							transaction,
-							errorCodes.PROCESSING_ERROR
+							apiCodes.PROCESSING_ERROR
 						);
 					})
 					.then(res => {
@@ -568,7 +568,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Application not found: ${unknownDappId}`
@@ -589,7 +589,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Application not found: ${inexistentId}`
@@ -609,7 +609,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Application not found: ${transactionsToWaitFor[0]}`
@@ -645,7 +645,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 				return sendTransactionPromise(
 					transaction,
-					errorCodes.PROCESSING_ERROR
+					apiCodes.PROCESSING_ERROR
 				).then(res => {
 					expect(res.body.message).to.match(
 						/^Account does not have enough LSK: /
@@ -701,7 +701,7 @@ describe('POST /api/transactions (type 7) outTransfer dapp', () => {
 
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction type ${transaction.type} is frozen`
