+++ framework/test/mocha/functional/http/post/transactions.js
@@ -21,7 +21,7 @@ const randomUtil = require('../../../common/utils/random');
 const accountFixtures = require('../../../fixtures/accounts');
 const sendTransactionPromise = require('../../../common/helpers/api')
 	.sendTransactionPromise;
-const errorCodes = require('../../../../../src/modules/http_api/helpers/api_codes');
+const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const phases = require('../../../common/phases');
 
 describe('POST /api/transactions (general)', () => {
@@ -68,7 +68,7 @@ describe('POST /api/transactions (general)', () => {
 		it('should fail when trying to send a transaction that is already confirmed', async () => {
 			return sendTransactionPromise(
 				transaction,
-				errorCodes.PROCESSING_ERROR
+				apiCodes.PROCESSING_ERROR
 			).then(res => {
 				expect(res.body.message).to.be.equal(
 					`Transaction is already confirmed: ${transaction.id}`
