+++ framework/test/mocha/unit/modules/chain/modules/multisignatures.js
@@ -18,8 +18,8 @@ const rewire = require('rewire');
 const accountsFixtures = require('../../../../fixtures/index').accounts;
 const transactionsFixtures = require('../../../../fixtures/index').transactions;
 const transactionTypes = require('../../../../../../src/modules/chain/helpers/transaction_types.js');
-const ApiError = require('../../../../../../src/modules/http_api/helpers/api_error');
-const errorCodes = require('../../../../../../src/modules/http_api/helpers/api_codes');
+const ApiError = require('../../../../../../src/modules/http_api/api_error');
+const apiCodes = require('../../../../../../src/modules/http_api/api_codes');
 
 const RewiredMultisignatures = rewire(
 	'../../../../../../src/modules/chain/modules/multisignatures.js'
@@ -1013,7 +1013,7 @@ describe('multisignatures', () => {
 			self.getGroup('', (err, scopeGroup) => {
 				expect(err).to.be.an.instanceof(ApiError);
 				expect(err.message).to.equal('Multisignature account not found');
-				expect(err.code).to.equal(errorCodes.NOT_FOUND);
+				expect(err.code).to.equal(apiCodes.NOT_FOUND);
 				expect(scopeGroup).to.not.exist;
 				done();
 			});
